# docs/guides/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Project Manager UI API
  description: |
    REST API for the Project Manager UI — a real-time project management dashboard
    for stakeholders and AI agents. Built with Next.js 15, PostgreSQL 16, Prisma, and SSE.

    **Base URL**: `http://localhost:3000/api` (local) or your deployment host.

    **Authentication** (dual):
    - **Agent access**: `X-API-Key` header (see [ApiKeyAuth](#/components/securitySchemes/ApiKeyAuth))
    - **Browser user access**: `auth_token` cookie set by login/register (see [CookieAuth](#/components/securitySchemes/CookieAuth))
    
    Write operations (POST, PUT, DELETE) require authentication. Read operations (GET) 
    are publicly accessible for backward compatibility (planned enforcement in v2).

    The SSE endpoint (`/api/events`) requires authentication as of v1.1.0.
  version: "1.1.0"
  contact:
    name: Project Manager UI Team
    url: https://github.com/gansru/ui_pm
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-domain.com/api
    description: Production server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for AI agent authentication. Valid for all authenticated operations.
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT cookie for browser user authentication. Set automatically by login/register endpoints. httpOnly, sameSite=lax, 7-day expiry.

    # ─── Request Payloads ───
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - confirmPassword
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Display name (required)
          example: "Roman Govorov"
        email:
          type: string
          format: email
          maxLength: 255
          description: Email address (must not already exist)
          example: "roman@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          description: Password (min 8 characters, cannot be a common password)
          example: "MyStr0ng!P@ss"
        confirmPassword:
          type: string
...
  schemas:
...
    ErrorBody:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorResponse"
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples:
            - VALIDATION_ERROR
            - NOT_FOUND
            - CONFLICT
            - UNAUTHORIZED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error description
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                description: Dot-path of the invalid field
              message:
                type: string
                description: Validation error message
          description: Detailed validation errors (present only for VALIDATION_ERROR)
...
paths:
...
  /auth/register:
    post:
      tags:
        - auth
      summary: Register a new user account
      description: |
        Creates a new user account with email/password authentication. Returns a JWT in an httpOnly cookie 
        for browser-session access. On successful registration, the user is automatically authenticated.
        
        **Anti-enumeration**: If the email already exists, returns `200` with a generic message (no cookie set).
        This prevents attackers from discovering which emails are registered.
        
        **Rate limiting**: Registration endpoint is rate-limited globally (100 req/min/IP).
        
        **Password requirements**: Min 8 characters, max 128 characters. Common passwords from known breaches are rejected.
        
        Requires `REGISTRATION_ENABLED=true` (default). Set to `'false'` to disable open registration.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              valid:
                summary: Valid registration
                value:
                  name: "Roman Govorov"
                  email: "roman@example.com"
                  password: "MyStr0ng!P@ss"
                  confirmPassword: "MyStr0ng!P@ss"
              weak_password:
                summary: Rejected common password
                value:
                  name: "Roman Govorov"
                  email: "roman@example.com"
                  password: "password123"
                  confirmPassword: "password123"
              mismatched:
                summary: Passwords do not match
                value:
                  name: "Roman Govorov"
                  email: "roman@example.com"
                  password: "MyStr0ng!P@ss"
                  confirmPassword: "Different123"
      responses:
        "201":
          description: Account created successfully. JWT cookie set for immediate login.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "auth_token=eyJ...; Path=/; Max-Age=604800; HttpOnly; SameSite=Lax"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
              example:
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Roman Govorov"
                  email: "roman@example.com"
                  role: stakeholder
        "200":
          description: |
            Email already exists — anti-enumeration response. No cookie is set and no session is created.
            The user should navigate to `/login` instead.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericMessage"
              example:
                message: "If an account with this email exists, you can sign in at /login"
        "400":
          description: Validation error — invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              examples:
                validation_error:
                  summary: Field validation failure
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid input
                      details:
                        - path: "email"
                          message: "Invalid email format"
                common_password:
                  summary: Password too common
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid input
                      details:
                        - path: "password"
                          message: "This password is too common. Please choose a stronger password."
                password_mismatch:
                  summary: Passwords do not match
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid input
                      details:
                        - path: "confirmPassword"
                          message: "Passwords do not match"
        "403":
          description: Registration is disabled (`REGISTRATION_ENABLED=false`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error:
                  code: REGISTRATION_DISABLED
                  message: "Registration is currently disabled"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user and create session
      description: |
        Authenticates a user with email/password. Returns a JWT in an httpOnly cookie that 
        is automatically sent with subsequent requests for 7 days.
        
        **Security notes**:
        - Rate limited on *failed* attempts only (successful logins don't consume quota): 10 failures/min/IP
        - Generic error message regardless of whether the email exists or the password is wrong
        - Audit-logged for security monitoring
        
        CORS: Credentials enabled when `origin` matches `CORS_ALLOWED_ORIGINS`. Clients must use 
        `credentials: 'include'` or `credentials: 'same-origin'` in fetch requests.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              valid:
                summary: Valid credentials
                value:
                  email: "roman@example.com"
                  password: "MyStr0ng!P@ss"
              empty_email:
                summary: Missing email
                value:
                  email: ""
                  password: "MyStr0ng!P@ss"
      responses:
        "200":
          description: Login successful. JWT cookie set.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "auth_token=eyJ...; Path=/; Max-Age=604800; HttpOnly; SameSite=Lax"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
              example:
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Roman Govorov"
                  email: "roman@example.com"
                  role: stakeholder
        "400":
          description: Validation error — invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid input
                  details:
                    - path: "email"
                      message: "Invalid email format"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
        "429":
          description: Too many failed login attempts. Wait and retry.
          headers:
            Retry-After:
              schema:
                type: integer
                description: Seconds until the rate limit window resets
                example: 60
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
              example:
                error:
                  code: RATE_LIMITED
                  message: "Too many failed login attempts. Please try again later."
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /auth/logout:
    post:
      tags:
        - auth
      summary: End current session
      description: |
        Clears the `auth_token` cookie, effectively ending the browser session. 
        Audit-logged if the user was authenticated.
        
        Best effort — if the API call fails, the local session state should be cleared 
        client-side regardless (the cookie expires on next page navigation anyway).
      operationId: logoutUser
      responses:
        "200":
          description: Logged out successfully. Cookie cleared.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "auth_token=; Path=/; Max-Age=0; HttpOnly; SameSite=Lax"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
              example:
                message: "Logged out successfully"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...